Got query module started, mostly as an example for me/xbony in the future when writing more methods
